Add piet::util module
This is intended for bits of code shared between various backend
implementations.

closes #183
